+++ UniFiTooling/UniFiTooling.psm1
@@ -5,6 +5,461 @@
 #endregion ModulePreLoaded
 
 #region ModulePrivateFunctions
+function Add-HostsEntry
+{
+   <#
+         .SYNOPSIS
+         Add a single Hosts Entry to the HOSTS File
+
+         .DESCRIPTION
+         Add a single Hosts Entry to the HOSTS File, multiple are not supported yet!
+
+         .PARAMETER Path
+         The path to the hosts file where the entry should be set. Defaults to the local computer's hosts file.
+
+         .PARAMETER Address
+         The Address address for the hosts entry.
+
+         .PARAMETER HostName
+         The hostname for the hosts entry.
+
+         .PARAMETER force
+         Force (replace)
+
+         .EXAMPLE
+         PS C:\> Add-HostsEntry -Address '0.0.0.0' -HostName 'badhost'
+
+         Add the host 'badhost' with the Adress '0.0.0.0' (blackhole) wo the Hosts.
+         If an Entry for 'badhost' exists, the new one will be appended anyway (You end up with two entries)
+
+         .EXAMPLE
+         PS C:\> Add-HostsEntry -Address '0.0.0.0' -HostName 'badhost' -force
+
+         Add the host 'badhost' with the Adress '0.0.0.0' (blackhole) wo the Hosts.
+         If an Entry for 'badhost' exists, the new one will replace the existing one.
+
+         .NOTES
+         Internal Helper, inspired by an old GIST I found
+
+         .LINK
+         Get-HostsFile
+
+         .LINK
+         Remove-HostsEntry
+
+         .LINK
+         https://gist.github.com/markembling/173887/1824b370be3fe468faceaed5f39b12bad010a417
+   #>
+
+   [CmdletBinding(ConfirmImpact = 'Medium',
+   SupportsShouldProcess)]
+   param
+   (
+      [Parameter(Mandatory,
+            Position = 0,
+      HelpMessage = 'The IP address for the hosts entry.')]
+      [ValidateNotNullOrEmpty()]
+      [Alias('ipaddress', 'ip')]
+      [string]
+      $Address,
+      [Parameter(Mandatory,
+            ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+            Position = 1,
+      HelpMessage = 'The hostname for the hosts entry.')]
+      [ValidateNotNullOrEmpty()]
+      [Alias('Host', 'Name')]
+      [string]
+      $HostName,
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 3)]
+      [switch]
+      $force = $false,
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 2)]
+      [ValidateNotNullOrEmpty()]
+      [Alias('filename', 'Hosts', 'hostsfile', 'file')]
+      [string]
+      $Path = "$env:windir\System32\drivers\etc\hosts"
+   )
+   begin {
+      Write-Verbose -Message 'Start'
+   }
+
+   process {
+      if ($force)
+      {
+         try
+         {
+            $null = (Remove-HostsEntry -HostName $HostName -Path $Path -ErrorAction SilentlyContinue -WarningAction SilentlyContinue)
+         }
+         catch
+         {
+            Write-Verbose -Message 'Looks like the entry was not there before'
+         }
+      }
+
+      try
+      {
+         if ($pscmdlet.ShouldProcess('Target', 'Operation'))
+         {
+            # Get a clean (end of) file
+            $paramGetContent = @{
+               Path          = $Path
+               Raw           = $true
+               Force         = $true
+               ErrorAction   = 'Stop'
+               WarningAction = 'SilentlyContinue'
+            }
+            $HostsFileContent = (((Get-Content @paramGetContent ).TrimEnd()).ToString())
+
+            $NewValue = "`n" + $Address + "`t`t" + $HostName
+            $NewHostsFileContent = $HostsFileContent + $NewValue
+
+            $paramSetContent = @{
+               Path          = $Path
+               Value         = $NewHostsFileContent
+               Force         = $true
+               Confirm       = $false
+               Encoding      = 'UTF8'
+               ErrorAction   = 'Stop'
+               WarningAction = 'SilentlyContinue'
+            }
+            $null = (Set-Content @paramSetContent)
+         }
+      }
+      catch
+      {
+         # get error record
+         [Management.Automation.ErrorRecord]$e = $_
+
+         # retrieve information about runtime error
+         $info = [PSCustomObject]@{
+            Exception = $e.Exception.Message
+            Reason    = $e.CategoryInfo.Reason
+            Target    = $e.CategoryInfo.TargetName
+            Script    = $e.InvocationInfo.ScriptName
+            Line      = $e.InvocationInfo.ScriptLineNumber
+            Column    = $e.InvocationInfo.OffsetInLine
+         }
+
+         Write-Verbose -Message $info
+
+         Write-Error -Message ($info.Exception) -ErrorAction Stop
+
+         # Only here to catch a global ErrorAction overwrite
+         break
+      }
+   }
+
+   end {
+      Write-Verbose -Message 'Done'
+   }
+}
+function ConvertFrom-UnixTimeStamp
+{
+   <#
+         .SYNOPSIS
+         Converts a Timestamp (Epochdate) into Datetime
+
+         .DESCRIPTION
+         Converts a Timestamp (Epochdate) into Datetime
+
+         .PARAMETER TimeStamp
+         Timestamp (Epochdate)
+
+         .PARAMETER Milliseconds
+         Is the given Timestamp (Epochdate) in Miliseconds instead of Seconds?
+
+         .EXAMPLE
+         PS C:\> ConvertFrom-UnixTimeStamp -TimeStamp 1547839380
+
+         Converts a Timestamp (Epochdate) into Datetime
+
+         .EXAMPLE
+         PS C:\> ConvertFrom-UnixTimeStamp -TimeStamp 1547839380712 -Milliseconds
+
+         Converts a Timestamp (Epochdate) into Datetime, given value is in Milliseconds
+
+         .NOTES
+         Added the 'UniFi' (Alias for the switch 'Milliseconds') because the API returns miliseconds instead of seconds
+   #>
+
+   [CmdletBinding(ConfirmImpact = 'None')]
+   [OutputType([datetime])]
+   param
+   (
+      [Parameter(Mandatory,
+            ValueFromPipeline,
+            Position = 0,
+      HelpMessage = 'Timestamp (Epochdate)')]
+      [ValidateNotNullOrEmpty()]
+      [Alias('Epochdate')]
+      [long]
+      $TimeStamp,
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 1)]
+      [Alias('UniFi')]
+      [switch]
+      $Milliseconds = $false
+   )
+
+   begin
+   {
+      # Set some defaults
+      $UnixStartTime = '1/1/1970'
+
+      # Cleanup
+      $Result = $null
+   }
+
+   process
+   {
+      try
+      {
+         if ($Milliseconds)
+         {
+            $Result = ((Get-Date -Date $UnixStartTime -ErrorAction Stop -WarningAction SilentlyContinue).AddMilliseconds($TimeStamp))
+         }
+         else
+         {
+            try
+            {
+               $Result = ((Get-Date -Date $UnixStartTime -ErrorAction Stop -WarningAction SilentlyContinue).AddSeconds($TimeStamp))
+            }
+            catch
+            {
+               # Try a Fallback!
+               $Result = ((Get-Date -Date $UnixStartTime -ErrorAction Stop -WarningAction SilentlyContinue).AddMilliseconds($TimeStamp))
+            }
+         }
+      }
+      catch
+      {
+         # get error record
+         [Management.Automation.ErrorRecord]$e = $_
+
+         # retrieve information about runtime error
+         $info = [PSCustomObject]@{
+            Exception = $e.Exception.Message
+            Reason    = $e.CategoryInfo.Reason
+            Target    = $e.CategoryInfo.TargetName
+            Script    = $e.InvocationInfo.ScriptName
+            Line      = $e.InvocationInfo.ScriptLineNumber
+            Column    = $e.InvocationInfo.OffsetInLine
+         }
+
+         Write-Verbose -Message $info
+
+         Write-Error -Message ($info.Exception) -ErrorAction Stop
+
+         # Only here to catch a global ErrorAction overwrite
+         break
+      }
+   }
+
+   end
+   {
+      $Result
+   }
+}
+function ConvertTo-UnixTimeStamp
+{
+   <#
+         .SYNOPSIS
+         Converts a Datetime into a Unix Timestamp (Epochdate)
+
+         .DESCRIPTION
+         Converts a Datetime into a Unix Timestamp (Epochdate)
+
+         .PARAMETER Date
+         The Date String that shoul be converted, default is now (if none is given)
+
+         .PARAMETER Milliseconds
+         Should the Timestamp (Epochdate) in Miliseconds instead of Seconds?
+
+         .EXAMPLE
+         PS C:\> ConvertTo-UnixTimeStamp
+
+         Converts the actual time into a Unix Timestamp (Epochdate)
+
+         .EXAMPLE
+         PS C:\> ConvertTo-UnixTimeStamp -Milliseconds
+
+         Converts the actual time into a Unix Timestamp (Epochdate), in milliseconds
+
+         .EXAMPLE
+         PS C:\> ConvertTo-UnixTimeStamp -Date ((Get-Date).AddDays(-1))
+
+         Covert the same time yesterday into a Unix Timestamp (Epochdate)
+
+         .EXAMPLE
+         PS C:\> ConvertTo-UnixTimeStamp -Date ((Get-Date).AddDays(-1)) -Milliseconds
+
+         Covert the same time yesterday into a Unix Timestamp (Epochdate), in milliseconds
+
+         .NOTES
+         Added the 'UniFi' (Alias for the switch 'Milliseconds') because the API returns miliseconds instead of seconds
+   #>
+
+   [CmdletBinding(ConfirmImpact = 'None')]
+   [OutputType([long])]
+   param
+   (
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 0)]
+      [ValidateNotNullOrEmpty()]
+      [Alias('TimeStamp', 'DateTimeStamp')]
+      [datetime]
+      $Date = (Get-Date),
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 1)]
+      [Alias('UniFi')]
+      [switch]
+      $Milliseconds = $false
+   )
+
+   begin
+   {
+      # Set some defaults
+      $UnixStartTime = '1/1/1970'
+
+      # Cleanup
+      $Result = $null
+   }
+
+   process
+   {
+      try
+      {
+         if ($Milliseconds)
+         {
+            $Result = ([long]((New-TimeSpan -Start (Get-Date -Date $UnixStartTime -ErrorAction Stop -WarningAction SilentlyContinue) -End (Get-Date -Date $Date -ErrorAction Stop -WarningAction SilentlyContinue) -ErrorAction Stop -WarningAction SilentlyContinue).TotalMilliseconds))
+         }
+         else
+         {
+            $Result = ([long]((New-TimeSpan -Start (Get-Date -Date $UnixStartTime -ErrorAction Stop -WarningAction SilentlyContinue) -End (Get-Date -Date $Date -ErrorAction Stop -WarningAction SilentlyContinue) -ErrorAction Stop -WarningAction SilentlyContinue).TotalSeconds))
+         }
+      }
+      catch
+      {
+         # get error record
+         [Management.Automation.ErrorRecord]$e = $_
+
+         # retrieve information about runtime error
+         $info = [PSCustomObject]@{
+            Exception = $e.Exception.Message
+            Reason    = $e.CategoryInfo.Reason
+            Target    = $e.CategoryInfo.TargetName
+            Script    = $e.InvocationInfo.ScriptName
+            Line      = $e.InvocationInfo.ScriptLineNumber
+            Column    = $e.InvocationInfo.OffsetInLine
+         }
+
+         Write-Verbose -Message $info
+
+         Write-Error -Message ($info.Exception) -ErrorAction Stop
+
+         # Only here to catch a global ErrorAction overwrite
+         break
+      }
+   }
+
+   end
+   {
+      $Result
+   }
+}
+function Get-HostsFile
+{
+   <#
+         .SYNOPSIS
+         Print the HOSTS File in a more clean format
+
+         .DESCRIPTION
+         Print the HOSTS File in a more clean format
+
+         .PARAMETER Path
+         The path to the hosts file where the entry should be set. Defaults to the local computer's hosts file.
+
+         .PARAMETER raw
+         Print raw Hosts File
+
+         .EXAMPLE
+         PS C:\> Get-HostsFile
+
+         Print the HOSTS File in a more clean format
+
+         .EXAMPLE
+         PS C:\> Get-HostsFile -raw
+
+         Print the HOSTS File in the regular format
+
+         .NOTES
+         Internal Helper, inspired by an old GIST I found
+
+         .LINK
+         Add-HostsEntry
+
+         .LINK
+         Remove-HostsEntry
+
+         .LINK
+         https://gist.github.com/markembling/173887/1824b370be3fe468faceaed5f39b12bad010a417
+   #>
+
+   [CmdletBinding(ConfirmImpact = 'None')]
+   param
+   (
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 0)]
+      [ValidateNotNullOrEmpty()]
+      [Alias('Hosts', 'hostsfile', 'file', 'filename')]
+      [string]
+      $Path = "$env:windir\System32\drivers\etc\hosts",
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 1)]
+      [Alias('plain')]
+      [switch]
+      $raw = $false
+   )
+
+   begin
+   {
+      $HostsFileContent = Get-Content -Path $Path
+   }
+
+   process
+   {
+      foreach ($line in $HostsFileContent)
+      {
+         if ($raw)
+         {
+            Write-Output -InputObject $line
+         }
+         else
+         {
+            $bits = [regex]::Split($line, '\t+')
+            if ($bits.count -eq 2)
+            {
+               [string]$HostsFileLine = $bits
+
+               if (-not ($HostsFileLine.StartsWith('#')))
+               {
+                  Write-Output -InputObject $HostsFileLine
+               }
+            }
+         }
+      }
+   }
+}
 function Get-UniFiConfig
 {
    <#
@@ -573,6 +1028,172 @@ function Invoke-UniFiCidrWorkaround
 	}
 }
 
+function Remove-HostsEntry
+{
+   <#
+         .SYNOPSIS
+         Removes a single Hosts Entry from the HOSTS File
+
+         .DESCRIPTION
+         Removes a single Hosts Entry from the HOSTS File, multiple are not supported yet!
+
+         .PARAMETER Path
+         The path to the hosts file where the entry should be set. Defaults to the local computer's hosts file.
+
+         .PARAMETER HostName
+         The hostname for the hosts entry.
+
+         .EXAMPLE
+         PS C:\> Remove-HostsEntry -HostName 'Dummy'
+
+         Remove the entry for the host 'Dummy' from the HOSTS File
+
+         .NOTES
+         Internal Helper, inspired by an old GIST I found
+
+         .LINK
+         Get-HostsFile
+
+         .LINK
+         Add-HostsEntry
+
+         .LINK
+         https://gist.github.com/markembling/173887/1824b370be3fe468faceaed5f39b12bad010a417
+   #>
+
+   [CmdletBinding(ConfirmImpact = 'Medium',
+   SupportsShouldProcess)]
+   param
+   (
+      [Parameter(Mandatory,
+            ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+            Position = 0,
+      HelpMessage = 'The hostname for the hosts entry.')]
+      [ValidateNotNullOrEmpty()]
+      [Alias('Host', 'Name')]
+      [string]
+      $HostName,
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 1)]
+      [ValidateNotNullOrEmpty()]
+      [Alias('Hosts', 'hostsfile', 'file', 'Filename')]
+      [string]
+      $Path = "$env:windir\System32\drivers\etc\hosts"
+   )
+
+   begin {
+      Write-Verbose -Message 'Start'
+
+      try
+      {
+         $paramGetContent = @{
+            Path          = $Path
+            Raw           = $true
+            Force         = $true
+            ErrorAction   = 'Stop'
+            WarningAction = 'SilentlyContinue'
+         }
+         $HostsFileContent = (((Get-Content @paramGetContent ).TrimEnd()).ToString())
+      }
+      catch
+      {
+         # get error record
+         [Management.Automation.ErrorRecord]$e = $_
+
+         # retrieve information about runtime error
+         $info = [PSCustomObject]@{
+            Exception = $e.Exception.Message
+            Reason    = $e.CategoryInfo.Reason
+            Target    = $e.CategoryInfo.TargetName
+            Script    = $e.InvocationInfo.ScriptName
+            Line      = $e.InvocationInfo.ScriptLineNumber
+            Column    = $e.InvocationInfo.OffsetInLine
+         }
+
+         Write-Verbose -Message $info
+
+         Write-Error -Message ($info.Exception) -ErrorAction Stop
+
+         # Only here to catch a global ErrorAction overwrite
+         break
+      }
+
+      $newLines = @()
+   }
+
+   process {
+      foreach ($line in $HostsFileContent)
+      {
+         $bits = [regex]::Split($line, '\t+')
+         if ($bits.count -eq 2)
+         {
+            if ($bits[1] -ne $HostName)
+            {
+               $newLines += $line
+            }
+         }
+         else
+         {
+            $newLines += $line
+         }
+      }
+
+      # Write file
+      try
+      {
+         if ($pscmdlet.ShouldProcess('Target', 'Operation'))
+         {
+            $paramClearContent = @{
+               Path          = $Path
+               Force         = $true
+               Confirm       = $false
+               ErrorAction   = 'Stop'
+               WarningAction = 'SilentlyContinue'
+            }
+            $null = (Clear-Content @paramClearContent)
+
+            $paramSetContent = @{
+               Path          = $Path
+               Value         = $newLines
+               Force         = $true
+               Confirm       = $false
+               Encoding      = 'UTF8'
+               ErrorAction   = 'Stop'
+               WarningAction = 'SilentlyContinue'
+            }
+            $null = (Set-Content @paramSetContent)
+         }
+      }
+      catch
+      {
+         # get error record
+         [Management.Automation.ErrorRecord]$e = $_
+
+         # retrieve information about runtime error
+         $info = [PSCustomObject]@{
+            Exception = $e.Exception.Message
+            Reason    = $e.CategoryInfo.Reason
+            Target    = $e.CategoryInfo.TargetName
+            Script    = $e.InvocationInfo.ScriptName
+            Line      = $e.InvocationInfo.ScriptLineNumber
+            Column    = $e.InvocationInfo.OffsetInLine
+         }
+
+         Write-Verbose -Message $info
+
+         Write-Error -Message ($info.Exception) -ErrorAction Stop
+
+         # Only here to catch a global ErrorAction overwrite
+         break
+      }
+   }
+
+   end {
+      Write-Verbose -Message 'Done'
+   }
+}
 function Set-UniFiApiLoginBody
       {
          <#